Removed Hanami::Action#accept? in favor of Hanami::Action::Request#accept?
